+++ NEWS.md
@@ -1,3 +1,8 @@
+# vehicletrends 0.0.4
+
+- Updated HHI data fixing error in 2025 year, and changed the name to `hhi_local`.
+- Added percentage of listings data at the census tract level and exported it as `p_local`.
+
 # vehicletrends 0.0.3
 
 - Added HHI data.